Add state management for cart
* Introduces an app-wide state for products on the cart and functions
  to add/remove products
* Adds an example implementation for a display of the cart on the
  navigation bar

We're using context, provided in the root layout, as there is no way
to pass custom props to pages in NextJS.

In order to use context, components have to be client components.
https://nextjs.org/docs/app/building-your-application/rendering/composition-patterns#using-context-providers
For this reason the `AddToCartButton` component is introduced. We still
want to fetch the product data on the server.